+++ trane/ops/transformation_ops.py
@@ -1,3 +1,5 @@
+import pandas as pd
+
 from trane.ops.op_base import OpBase
 
 
@@ -11,6 +13,19 @@ class TransformationOpBase(OpBase):
     operations are defined as classes that inherit the
     TransformationOpBase class and instantiate the execute method.
     """
+
+
+class IdentityOp(TransformationOpBase):
+    input_output_types = [("None", "None")]
+    description = ""
+
+    def generate_description(self):
+        return self.description
+
+    def label_function(self, dataslice):
+        if len(dataslice) == 0:
+            return pd.NA
+        return dataslice
 
 
 class OrderByOp(TransformationOpBase):